Introduce RPC module
